requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ML and NLP libraries

spacy>=3.8.0
transformers==4.35.0


# HTTP and file handling
python-multipart==0.0.6
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1

numpy>=1.19.0